+++ KvotaWeb/Models/Items/Tampopechat.cs
@@ -47,10 +47,10 @@ namespace KvotaWeb.Models.Items
                 ret.Add(line);
                 if (KolichestvoTcvetov == null || Tiraz == null) continue;
 
-                double cena;
+                decimal cena;
                 if (TryGetPrice(i, Tiraz, KolichestvoTcvetov, out cena) == false) continue;
 
-                line.Cena = cena * Tiraz.Value;
+                line.Cena = cena * (decimal)Tiraz.Value;
             }
             return ret;
